+++ null-log/src/arguments/parser.hh
@@ -1,9 +1,10 @@
 #ifndef PARSER_H
 #define PARSER_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 // Parse Command Line Arguments
-int8_t parse_arguments(int length, char *arguments[]);
+uint8_t parse_arguments(size_t length, char *arguments[]);
 
 #endif
